dispatcher/requirements.txt
django
south
django_extensions
# libyaml must be installed too
pyyaml

